# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: biyaHERO Transportation Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List all registered vehicles
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/VehicleType"
          description: Filter by vehicle type
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, en_route]
          description: Filter by vehicle status
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get a single vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    VehicleType:
      type: string
      enum: [jeepney, tricycle, bus, van, fx, uv_express, ferry]
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        type:
          $ref: "#/components/schemas/VehicleType"
        plateNumber:
          type: string
        operator:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [active, inactive, en_route]
        currentLat:
          type: number
          format: double
        currentLng:
          type: number
          format: double
        routeId:
          type: integer
          nullable: true
        routeName:
          type: string
          nullable: true
      required:
        - id
        - type
        - plateNumber
        - operator
        - capacity
        - status
        - currentLat
        - currentLng
...